Refactor make_config as it is getting unweildy
