+++ ci/prow/make_config.go
@@ -40,18 +40,23 @@ import (
 
 const (
 	// Manifests generated by ko are indented by 2 spaces.
-	baseIndent = "  "
-	// Cron strings for key jobs
-	goCoveragePeriodicJobCron        = "0 1 * * *"  // Run at 01:00 every day
-	cleanupPeriodicJobCron           = "0 19 * * 1" // Run at 11:00PST/12:00PST every Monday (19:00 UTC)
-	flakesReporterPeriodicJobCron    = "0 12 * * *" // Run at 4:00PST/5:00PST every day (12:00 UTC)
-	prowversionbumperPeriodicJobCron = "0 20 * * 1" // Run at 12:00PST/13:00PST every Monday (20:00 UTC)
-	backupPeriodicJobCron            = "15 9 * * *" // Run at 02:15PST every day (09:15 UTC)
+	baseIndent  = "  "
+	templateDir = "templates"
 
-	// baseOptions setting for testgrid dashboard tabs
-	testgridTabGroupByDir    = "exclude-filter-by-regex=Overall$&group-by-directory=&expand-groups=&sort-by-name="
-	testgridTabGroupByTarget = "exclude-filter-by-regex=Overall$&group-by-target=&expand-groups=&sort-by-name="
-	testgridTabSortByName    = "sort-by-name="
+	// ##########################################################
+	// ############## prow configuration templates ##############
+	// ##########################################################
+	// generalProwConfig contains config-wide definitions.
+	generalProwConfig = "prow_config_header.yaml"
+
+	// presubmitJob is the template for presubmit jobs.
+	presubmitJob = "prow_presubmit_job.yaml"
+
+	// presubmitGoCoverageJob is the template for go coverage presubmit jobs.
+	presubmitGoCoverageJob = "prow_presubmit_gocoverate_job.yaml"
+
+	// goCoveragePostsubmitJob is the template for the go postsubmit coverage job.
+	goCoveragePostsubmitJob = "prow_postsubmit_gocoverage_job.yaml"
 )
 
 // repositoryData contains basic data about each Knative repository.
@@ -108,14 +113,6 @@ type presubmitJobTemplateData struct {
 	PresubmitCommand     []string
 }
 
-// periodicJobTemplateData contains data about a periodic Prow job.
-type periodicJobTemplateData struct {
-	Base            baseProwJobTemplateData
-	PeriodicJobName string
-	CronString      string
-	PeriodicCommand []string
-}
-
 // postsubmitJobTemplateData contains data about a postsubmit Prow job.
 type postsubmitJobTemplateData struct {
 	Base              baseProwJobTemplateData
@@ -130,41 +127,6 @@ type newJobNeeded func(repositoryData) bool
 
 // stringArrayFlag is the content of a multi-value flag.
 type stringArrayFlag []string
-
-// ####################################################################################################
-// ############## data definitions that are used for the testgrid config file generation ##############
-// ####################################################################################################
-// baseTestgridTemplateData contains basic data about the testgrid config file.
-// TODO(Fredy-Z): remove this structure and use baseProwJobTemplateData instead
-type baseTestgridTemplateData struct {
-	TestGroupName string
-	Year          int
-}
-
-// testGroupTemplateData contains data about a test group
-type testGroupTemplateData struct {
-	Base baseTestgridTemplateData
-	// TODO(Fredy-Z): use baseProwJobTemplateData then this attribute can be removed
-	GcsLogDir string
-	Extras    map[string]string
-}
-
-// dashboardTabTemplateData contains data about a dashboard tab
-type dashboardTabTemplateData struct {
-	Base        baseTestgridTemplateData
-	Name        string
-	BaseOptions string
-	Extras      map[string]string
-}
-
-// dashboardGroupTemplateData contains data about a dashboard group
-type dashboardGroupTemplateData struct {
-	Name      string
-	RepoNames []string
-}
-
-// testgridEntityGenerator is a function that generates the entity given the repo name and job names
-type testgridEntityGenerator func(string, string, []string)
 
 var (
 	// Values used in the jobs that can be changed through command-line flags.
@@ -205,67 +167,6 @@ var (
 
 	// Map which sections of the config.yaml were written to stdout.
 	sectionMap map[string]bool
-
-	// #############################################################################
-	// ############## data used for generating testgrid configuration ##############
-	// #############################################################################
-	// goCoverageMap keep track of which repo has go code coverage when parsing the simple config file
-	goCoverageMap map[string]bool
-	// projNames save the proj names in a list when parsing the simple config file, for the purpose of maintaining the output sequence
-	projNames []string
-	// repoNames save the repo names in a list when parsing the simple config file, for the purpose of maintaining the output sequence
-	repoNames []string
-
-	// metaData saves the meta data needed to generate the final config file
-	// key is the main project version, value is another map containing job details
-	//     for the job detail map, key is the repo name, value is the list of job types, like continuous, latency, nightly, and etc.
-	metaData = make(map[string]map[string][]string)
-
-	// templatesCache caches templates in memory to avoid I/O
-	templatesCache = make(map[string]string)
-)
-
-// Templates for config generation.
-
-const (
-	templateDir = "templates"
-
-	// ##########################################################
-	// ############## prow configuration templates ##############
-	// ##########################################################
-	// generalProwConfig contains config-wide definitions.
-	generalProwConfig = "prow_config_header.yaml"
-
-	// presubmitJob is the template for presubmit jobs.
-	presubmitJob = "prow_presubmit_job.yaml"
-
-	// presubmitGoCoverageJob is the template for go coverage presubmit jobs.
-	presubmitGoCoverageJob = "prow_presubmit_gocoverate_job.yaml"
-
-	// periodicTestJob is the template for periodic test/release jobs.
-	periodicTestJob = "prow_periodic_test_job.yaml"
-
-	// periodicCustomJob is the template for periodic custom jobs.
-	periodicCustomJob = "prow_periodic_custom_job.yaml"
-
-	// goCoveragePostsubmitJob is the template for the go postsubmit coverage job.
-	goCoveragePostsubmitJob = "prow_postsubmit_gocoverage_job.yaml"
-
-	// ##############################################################
-	// ############## testgrid configuration templates ##############
-	// ##############################################################
-
-	// generalTestgridConfig contains config-wide definitions.
-	generalTestgridConfig = "testgrid_config_header.yaml"
-
-	// testGroupTemplate is the template for the test group config
-	testGroupTemplate = "testgrid_testgroup.yaml"
-
-	// dashboardTabTemplate is the template for the dashboard tab config
-	dashboardTabTemplate = "testgrid_dashboardtab.yaml"
-
-	// dashboardGroupTemplate is the template for the dashboard tab config
-	dashboardGroupTemplate = "testgrid_dashboardgroup.yaml"
 )
 
 // Generate cron string based on job type, offset generated from jobname
@@ -424,14 +325,6 @@ func newbaseProwJobTemplateData(repo string) baseProwJobTemplateData {
 	data.Env = make([]string, 0)
 	data.ExtraRefs = []string{"- org: " + data.OrgName, "  repo: " + data.RepoName, "  base_ref: master"}
 	data.Labels = make([]string, 0)
-	return data
-}
-
-// newBaseTestgridTemplateData returns a testgridTemplateData type with its initial, default values.
-func newBaseTestgridTemplateData(testGroupName string) baseTestgridTemplateData {
-	var data baseTestgridTemplateData
-	data.Year = time.Now().Year()
-	data.TestGroupName = testGroupName
 	return data
 }
 
@@ -611,250 +504,6 @@ func generatePresubmit(title string, repoName string, presubmitConfig yaml.MapSl
 		data.Base.Image = strings.Replace(data.Base.Image, "coverage:latest", "coverage-dev:latest-dev", -1)
 		template := strings.Replace(readTemplate(presubmitGoCoverageJob), "(all|", "(", 1)
 		executeJobTemplate("presubmit", template, title, repoName, data.PresubmitPullJobName, true, data)
-	}
-}
-
-// generatePeriodic generates all periodic job configs for the given repo and configuration.
-func generatePeriodic(title string, repoName string, periodicConfig yaml.MapSlice) {
-	var data periodicJobTemplateData
-	data.Base = newbaseProwJobTemplateData(repoName)
-	jobNameSuffix := ""
-	jobTemplate := readTemplate(periodicTestJob)
-	jobType := ""
-
-	for i, item := range periodicConfig {
-		switch item.Key {
-		case "continuous":
-			if !getBool(item.Value) {
-				return
-			}
-			jobType = getString(item.Key)
-			jobNameSuffix = "continuous"
-			// Use default command and arguments if none given.
-			if data.Base.Command == "" {
-				data.Base.Command = presubmitScript
-			}
-			if len(data.Base.Args) == 0 {
-				data.Base.Args = allPresubmitTests
-			}
-		case "nightly":
-			if !getBool(item.Value) {
-				return
-			}
-			jobType = getString(item.Key)
-			jobNameSuffix = "nightly-release"
-			data.Base.ServiceAccount = nightlyAccount
-			data.Base.Command = releaseScript
-			data.Base.Args = releaseNightly
-			data.Base.Timeout = 90
-		case "branch-ci":
-			if !getBool(item.Value) {
-				return
-			}
-			jobType = getString(item.Key)
-			jobNameSuffix = "continuous"
-			data.Base.Command = releaseScript
-			data.Base.Args = releaseLocal
-			setupDockerInDockerForJob(&data.Base)
-			// TODO(adrcunha): Consider reducing the timeout in the future.
-			data.Base.Timeout = 180
-		case "dot-release", "auto-release":
-			if !getBool(item.Value) {
-				return
-			}
-			jobType = getString(item.Key)
-			jobNameSuffix = getString(item.Key)
-			data.Base.ServiceAccount = releaseAccount
-			data.Base.Command = releaseScript
-			data.Base.Args = []string{
-				"--" + jobNameSuffix,
-				"--release-gcs " + data.Base.ReleaseGcs,
-				"--release-gcr gcr.io/knative-releases",
-				"--github-token /etc/hub-token/token"}
-			addVolumeToJob(&data.Base, "/etc/hub-token", "hub-token", true)
-			data.Base.Timeout = 90
-		case "performance", "performance-mesh":
-			if !getBool(item.Value) {
-				return
-			}
-			jobType = getString(item.Key)
-			jobNameSuffix = getString(item.Key)
-			data.Base.Command = performanceScript
-			// We need a larger cluster of at least 16 nodes for perf tests
-			addEnvToJob(&data.Base, "E2E_MIN_CLUSTER_NODES", "16")
-			addEnvToJob(&data.Base, "E2E_MAX_CLUSTER_NODES", "16")
-			addVolumeToJob(&data.Base, "/secrets/cloudsql/monitoringdb", "monitoring-db-credentials", true)
-			data.Base.Timeout = 120
-		case "latency":
-			if !getBool(item.Value) {
-				return
-			}
-			jobType = getString(item.Key)
-			jobTemplate = readTemplate(periodicCustomJob)
-			jobNameSuffix = "latency"
-			data.Base.Image = "gcr.io/knative-tests/test-infra/metrics:latest"
-			data.Base.Command = "/metrics"
-			data.Base.Args = []string{
-				fmt.Sprintf("--source-directory=ci-%s-continuous", data.Base.RepoNameForJob),
-				"--artifacts-dir=$(ARTIFACTS)",
-				"--service-account=" + data.Base.ServiceAccount}
-		case "custom-job":
-			jobType = getString(item.Key)
-			jobNameSuffix = getString(item.Value)
-			data.Base.Timeout = 100
-		case "cron":
-			data.CronString = getString(item.Value)
-		case "release":
-			version := getString(item.Value)
-			jobNameSuffix = version + "-" + jobNameSuffix
-			data.Base.RepoBranch = "release-" + version
-		case "webhook-apicoverage":
-			if !getBool(item.Value) {
-				return
-			}
-			jobType = getString(item.Key)
-			jobNameSuffix = "webhook-apicoverage"
-			data.Base.Command = webhookAPICoverageScript
-			addEnvToJob(&data.Base, "SYSTEM_NAMESPACE", data.Base.RepoNameForJob)
-		default:
-			continue
-		}
-		// Knock-out the item, signalling it was already parsed.
-		periodicConfig[i] = yaml.MapItem{}
-	}
-
-	parseBasicJobConfigOverrides(&data.Base, periodicConfig)
-	data.PeriodicJobName = fmt.Sprintf("ci-%s", data.Base.RepoNameForJob)
-	if jobNameSuffix != "" {
-		data.PeriodicJobName += "-" + jobNameSuffix
-	}
-	if data.CronString == "" {
-		data.CronString = generateCron(jobType, data.PeriodicJobName, data.Base.Timeout)
-	}
-	// Ensure required data exist.
-	if data.CronString == "" {
-		log.Fatalf("Job %q is missing cron string", data.PeriodicJobName)
-	}
-	if len(data.Base.Args) == 0 && data.Base.Command == "" {
-		log.Fatalf("Job %q is missing command", data.PeriodicJobName)
-	}
-	if jobType == "branch-ci" && data.Base.RepoBranch == "" {
-		log.Fatalf("%q jobs are intended to be used on release branches", jobType)
-	}
-	// Generate config itself.
-	data.PeriodicCommand = createCommand(data.Base)
-	if data.Base.ServiceAccount != "" {
-		addEnvToJob(&data.Base, "GOOGLE_APPLICATION_CREDENTIALS", data.Base.ServiceAccount)
-		addEnvToJob(&data.Base, "E2E_CLUSTER_REGION", "us-central1")
-	}
-	if data.Base.RepoBranch != "" {
-		// If it's a release version, add env var PULL_BASE_REF as ref name of the base branch.
-		// TODO(Fredy-Z): this serves as a workaround, see https://github.com/knative/test-infra/issues/780.
-		addEnvToJob(&data.Base, "PULL_BASE_REF", data.Base.RepoBranch)
-	}
-	addExtraEnvVarsToJob(&data.Base)
-	configureServiceAccountForJob(&data.Base)
-	executeJobTemplate("periodic", jobTemplate, title, repoName, data.PeriodicJobName, false, data)
-}
-
-// generateCleanupPeriodicJob generates the cleanup job config.
-func generateCleanupPeriodicJob() {
-	var data periodicJobTemplateData
-	data.Base = newbaseProwJobTemplateData("knative/test-infra")
-	data.PeriodicJobName = "ci-knative-cleanup"
-	data.CronString = cleanupPeriodicJobCron
-	data.Base.DecorationConfig = []string{"timeout: 86400000000000"} // 24 hours
-	data.Base.Command = cleanupScript
-	data.Base.Args = []string{
-		"--project-resource-yaml ci/prow/boskos/resources.yaml",
-		"--days-to-keep-images 30",
-		"--hours-to-keep-clusters 24",
-		"--service-account " + data.Base.ServiceAccount,
-		"--artifacts $(ARTIFACTS)"}
-	addExtraEnvVarsToJob(&data.Base)
-	configureServiceAccountForJob(&data.Base)
-	executeJobTemplate("periodic cleanup", readTemplate(periodicCustomJob), "presubmits", "", data.PeriodicJobName, false, data)
-}
-
-// generateFlakytoolPeriodicJob generates the cleanup job config.
-func generateFlakytoolPeriodicJob() {
-	var data periodicJobTemplateData
-	data.Base = newbaseProwJobTemplateData("knative/test-infra")
-	data.Base.Image = flakesreporterDockerImage
-	data.PeriodicJobName = "ci-knative-flakes-reporter"
-	data.CronString = flakesReporterPeriodicJobCron
-	data.Base.Command = "/flaky-test-reporter"
-	data.Base.Args = []string{
-		"--service-account=" + data.Base.ServiceAccount,
-		"--github-account=/etc/flaky-test-reporter-github-token/token",
-		"--slack-account=/etc/flaky-test-reporter-slack-token/token"}
-	addExtraEnvVarsToJob(&data.Base)
-	configureServiceAccountForJob(&data.Base)
-	addVolumeToJob(&data.Base, "/etc/flaky-test-reporter-github-token", "flaky-test-reporter-github-token", true)
-	addVolumeToJob(&data.Base, "/etc/flaky-test-reporter-slack-token", "flaky-test-reporter-slack-token", true)
-	executeJobTemplate("periodic flakesreporter", readTemplate(periodicCustomJob), "presubmits", "", data.PeriodicJobName, false, data)
-}
-
-// generateVersionBumpertoolPeriodicJob generates the Prow version bumper job config.
-func generateVersionBumpertoolPeriodicJob() {
-	var data periodicJobTemplateData
-	data.Base = newbaseProwJobTemplateData("knative/test-infra")
-	data.Base.Image = prowversionbumperDockerImage
-	data.PeriodicJobName = "ci-knative-prow-auto-bumper"
-	data.CronString = prowversionbumperPeriodicJobCron
-	data.Base.Command = "/prow-auto-bumper"
-	data.Base.Args = []string{
-		"--github-account=/etc/prow-auto-bumper-github-token/token",
-		"--git-userid=knative-prow-updater-robot",
-		"--git-username='Knative Prow Updater Robot'",
-		"--git-email=knative-prow-updater-robot@google.com"}
-	addExtraEnvVarsToJob(&data.Base)
-	configureServiceAccountForJob(&data.Base)
-	addVolumeToJob(&data.Base, "/etc/prow-auto-bumper-github-token", "prow-auto-bumper-github-token", true)
-	executeJobTemplate("periodic versionbumper", readTemplate(periodicCustomJob), "presubmits", "", data.PeriodicJobName, false, data)
-}
-
-// generateBackupPeriodicJob generates the backup job config.
-func generateBackupPeriodicJob() {
-	var data periodicJobTemplateData
-	data.Base = newbaseProwJobTemplateData("none/unused")
-	data.Base.ServiceAccount = "/etc/backup-account/service-account.json"
-	data.Base.Image = "gcr.io/knative-tests/test-infra/backups:latest"
-	data.PeriodicJobName = "ci-knative-backup-artifacts"
-	data.CronString = backupPeriodicJobCron
-	data.Base.Command = "/backup.sh"
-	data.Base.Args = []string{data.Base.ServiceAccount}
-	data.Base.ExtraRefs = []string{} // no repo clone required
-	addExtraEnvVarsToJob(&data.Base)
-	configureServiceAccountForJob(&data.Base)
-	executeJobTemplate("periodic backup", readTemplate(periodicCustomJob), "presubmits", "", data.PeriodicJobName, false, data)
-}
-
-// generateGoCoveragePeriodic generates the go coverage periodic job config for the given repo (configuration is ignored).
-func generateGoCoveragePeriodic(title string, repoName string, _ yaml.MapSlice) {
-	for i, repo := range repositories {
-		if repoName != repo.Name || !repo.EnableGoCoverage {
-			continue
-		}
-		repositories[i].Processed = true
-		var data periodicJobTemplateData
-		data.Base = newbaseProwJobTemplateData(repoName)
-		data.Base.Image = coverageDockerImage
-		data.PeriodicJobName = fmt.Sprintf("ci-%s-go-coverage", data.Base.RepoNameForJob)
-		data.CronString = goCoveragePeriodicJobCron
-		data.Base.GoCoverageThreshold = repo.GoCoverageThreshold
-		data.Base.Command = "/coverage"
-		data.Base.Args = []string{
-			"--artifacts=$(ARTIFACTS)",
-			fmt.Sprintf("--cov-threshold-percentage=%d", data.Base.GoCoverageThreshold)}
-		data.Base.ServiceAccount = ""
-		addExtraEnvVarsToJob(&data.Base)
-		addLabelToJob(&data.Base, "prow.k8s.io/pubsub.project", "knative-tests")
-		addLabelToJob(&data.Base, "prow.k8s.io/pubsub.topic", "knative-monitoring")
-		addLabelToJob(&data.Base, "prow.k8s.io/pubsub.runID", data.PeriodicJobName)
-		configureServiceAccountForJob(&data.Base)
-		executeJobTemplate("periodic go coverage", readTemplate(periodicCustomJob), title, repoName, data.PeriodicJobName, false, data)
-		return
 	}
 }
 
@@ -1189,141 +838,6 @@ func addRemainingTestCoverageJobs() {
 	}
 }
 
-// generateTestGridSection generates the configs for a TestGrid section using the given generator
-func generateTestGridSection(sectionName string, generator testgridEntityGenerator, skipReleasedProj bool) {
-	outputConfig(sectionName + ":")
-	for _, projName := range projNames {
-		// Do not handle the project if it is released and we want to skip it.
-		if skipReleasedProj && isReleased(projName) {
-			continue
-		}
-		repos := metaData[projName]
-		for _, repoName := range repoNames {
-			if jobNames, exists := repos[repoName]; exists {
-				generator(projName, repoName, jobNames)
-			}
-		}
-	}
-}
-
-// generateTestGroup generates the test group configuration
-func generateTestGroup(projName string, repoName string, jobNames []string) {
-	projRepoStr := buildProjRepoStr(projName, repoName)
-	for _, jobName := range jobNames {
-		testGroupName := getTestGroupName(projRepoStr, jobName)
-		gcsLogDir := fmt.Sprintf("%s/%s/%s", gcsBucket, logsDir, testGroupName)
-		extras := make(map[string]string)
-		switch jobName {
-		case "continuous", "dot-release", "auto-release", "performance", "performance-mesh", "latency", "nightly":
-			isDailyBranch := regexp.MustCompile(`-[0-9\.]+-continuous`).FindString(testGroupName) != ""
-			if !isDailyBranch && (jobName == "continuous" || jobName == "auto-release") {
-				// TODO(Fredy-Z): this value should be derived from the cron string
-				extras["alert_stale_results_hours"] = "3"
-				if jobName == "continuous" {
-					// For continuous flows, alert after 3 failures due to flakiness
-					extras["num_failures_to_alert"] = "3"
-				}
-			}
-			if jobName == "dot-release" {
-				// TODO(Fredy-Z): this value should be derived from the cron string
-				extras["alert_stale_results_hours"] = "170" // 1 week + 2h
-			}
-			if jobName == "latency" {
-				extras["short_text_metric"] = "latency"
-			}
-			if jobName == "performance" || jobName == "performance-mesh" {
-				extras["short_text_metric"] = "perf_latency"
-			}
-		case "test-coverage":
-			gcsLogDir = strings.ToLower(fmt.Sprintf("%s/%s/ci-%s-%s", gcsBucket, logsDir, projRepoStr, "go-coverage"))
-			extras["short_text_metric"] = "coverage"
-			// Do not alert on coverage failures (i.e., coverage below threshold)
-			extras["num_failures_to_alert"] = "9999"
-		case "istio-1.0.7-mesh", "istio-1.0.7-no-mesh", "istio-1.1.7-mesh", "istio-1.1.7-no-mesh",
-			"k8s-1.12-istio-1.1", "k8s-1.12-istio-1.0", "k8s-1.11-istio-1.1", "k8s-1.11-istio-1.0":
-			extras["alert_stale_results_hours"] = "3"
-			extras["num_failures_to_alert"] = "3"
-		default:
-			log.Fatalf("Unknown jobName for generateTestGroup: %s", jobName)
-		}
-		executeTestGroupTemplate(testGroupName, gcsLogDir, extras)
-	}
-}
-
-// executeTestGroupTemplate outputs the given test group config template with the given data
-func executeTestGroupTemplate(testGroupName string, gcsLogDir string, extras map[string]string) {
-	var data testGroupTemplateData
-	data.Base.TestGroupName = testGroupName
-	data.GcsLogDir = gcsLogDir
-	data.Extras = extras
-	executeTemplate("test group", readTemplate(testGroupTemplate), data)
-}
-
-// generateDashboard generates the dashboard configuration
-func generateDashboard(projName string, repoName string, jobNames []string) {
-	projRepoStr := buildProjRepoStr(projName, repoName)
-	outputConfig("- name: " + strings.ToLower(repoName) + "\n" + baseIndent + "dashboard_tab:")
-	noExtras := make(map[string]string)
-	for _, jobName := range jobNames {
-		testGroupName := getTestGroupName(projRepoStr, jobName)
-		switch jobName {
-		case "continuous":
-			executeDashboardTabTemplate("continuous", testGroupName, testgridTabSortByName, noExtras)
-			// This is a special case for knative/serving, as conformance tab is just a filtered view of the continuous tab.
-			if projRepoStr == "knative-serving" {
-				executeDashboardTabTemplate("conformance", testGroupName, "include-filter-by-regex=test/conformance/&sort-by-name=", noExtras)
-			}
-		case "dot-release", "auto-release", "performance", "performance-mesh", "latency":
-			extras := make(map[string]string)
-			baseOptions := testgridTabSortByName
-			if jobName == "performance" || jobName == "performance-mesh" {
-				baseOptions = testgridTabGroupByTarget
-			}
-			if jobName == "latency" {
-				baseOptions = testgridTabGroupByDir
-				extras["description"] = "95% latency in ms"
-			}
-			executeDashboardTabTemplate(jobName, testGroupName, baseOptions, extras)
-		case "nightly":
-			executeDashboardTabTemplate("nightly", testGroupName, testgridTabSortByName, noExtras)
-		case "test-coverage":
-			executeDashboardTabTemplate("coverage", testGroupName, testgridTabGroupByDir, noExtras)
-		case "istio-1.0.7-mesh", "istio-1.0.7-no-mesh", "istio-1.1.7-mesh", "istio-1.1.7-no-mesh",
-			"k8s-1.12-istio-1.1", "k8s-1.12-istio-1.0", "k8s-1.11-istio-1.1", "k8s-1.11-istio-1.0":
-			executeDashboardTabTemplate(jobName, testGroupName, testgridTabSortByName, noExtras)
-		default:
-			log.Fatalf("Unknown job name %q", jobName)
-		}
-	}
-}
-
-// executeTestGroupTemplate outputs the given dashboard tab config template with the given data
-func executeDashboardTabTemplate(dashboardTabName string, testGroupName string, baseOptions string, extras map[string]string) {
-	var data dashboardTabTemplateData
-	data.Name = dashboardTabName
-	data.Base.TestGroupName = testGroupName
-	data.BaseOptions = baseOptions
-	data.Extras = extras
-	executeTemplate("dashboard tab", readTemplate(dashboardTabTemplate), data)
-}
-
-// getTestGroupName get the testGroupName from the given repoName and jobName
-func getTestGroupName(repoName string, jobName string) string {
-	switch jobName {
-	case "continuous", "dot-release", "auto-release", "performance", "performance-mesh", "latency":
-		return strings.ToLower(fmt.Sprintf("ci-%s-%s", repoName, jobName))
-	case "nightly":
-		return strings.ToLower(fmt.Sprintf("ci-%s-%s-release", repoName, jobName))
-	case "test-coverage":
-		return strings.ToLower(fmt.Sprintf("pull-%s-%s", repoName, jobName))
-	case "istio-1.0.7-mesh", "istio-1.0.7-no-mesh", "istio-1.1.7-mesh", "istio-1.1.7-no-mesh",
-		"k8s-1.12-istio-1.1", "k8s-1.12-istio-1.0", "k8s-1.11-istio-1.1", "k8s-1.11-istio-1.0":
-		return strings.ToLower(fmt.Sprintf("ci-%s-%s", repoName, jobName))
-	}
-	log.Fatalf("Unknown jobName for getTestGroupName: %s", jobName)
-	return ""
-}
-
 // buildProjRepoStr builds the projRepoStr used in the config file with projName and repoName
 func buildProjRepoStr(projName string, repoName string) string {
 	projVersion := ""
@@ -1338,52 +852,6 @@ func buildProjRepoStr(projName string, repoName string) string {
 	}
 	projRepoStr = projName + "-" + projRepoStr
 	return strings.ToLower(projRepoStr)
-}
-
-func generateDashboardsForReleases() {
-	for _, projName := range projNames {
-		// Do not handle the project if it is not released.
-		if !isReleased(projName) {
-			continue
-		}
-		repos := metaData[projName]
-		outputConfig("- name: " + projName + "\n" + baseIndent + "dashboard_tab:")
-		noExtras := make(map[string]string)
-		for _, repoName := range repoNames {
-			if _, exists := repos[repoName]; exists {
-				testGroupName := getTestGroupName(buildProjRepoStr(projName, repoName), "continuous")
-				executeDashboardTabTemplate(repoName, testGroupName, testgridTabSortByName, noExtras)
-			}
-		}
-	}
-}
-
-// generateDashboardGroups generates the dashboard groups configuration
-func generateDashboardGroups() {
-	outputConfig("dashboard_groups:")
-	for _, projName := range projNames {
-		// there is only one dashboard for each released project, so we do not need to group them
-		if isReleased(projName) {
-			continue
-		}
-
-		dashboardRepoNames := make([]string, 0)
-		repos := metaData[projName]
-		for _, repoName := range repoNames {
-			if _, exists := repos[repoName]; exists {
-				dashboardRepoNames = append(dashboardRepoNames, repoName)
-			}
-		}
-		executeDashboardGroupTemplate(projName, dashboardRepoNames)
-	}
-}
-
-// executeDashboardGroupTemplate outputs the given dashboard group config template with the given data
-func executeDashboardGroupTemplate(dashboardGroupName string, dashboardRepoNames []string) {
-	var data dashboardGroupTemplateData
-	data.Name = dashboardGroupName
-	data.RepoNames = dashboardRepoNames
-	executeTemplate("dashboard group", readTemplate(dashboardGroupTemplate), data)
 }
 
 // isReleased returns true for project name that has version
